DEMO MODE ON

+++ src/main/java/com/medteamb/medteamb/controller/DoctorController.java
@@ -39,7 +39,7 @@ public class DoctorController {
 	}
 	
 	@GetMapping(path ="/get/{doctorId}")
-	public ResponseEntity<DoctorRequestDTO>showById(@PathVariable Integer doctorId){
+	public ResponseEntity<DoctorRequestDTO>showById(@PathVariable Long doctorId){
 		Optional<DoctorRequestDTO> foundDoctor = docService.findDocById(doctorId);
 		if(foundDoctor.isPresent()) {
 			DoctorRequestDTO doctorDto = foundDoctor.get();
@@ -50,7 +50,7 @@ public class DoctorController {
 	}
 	
 	@PutMapping(path = "/update/{doctorId}")
-	public ResponseEntity<DoctorRequestDTO>updateById(@PathVariable Integer doctorId,
+	public ResponseEntity<DoctorRequestDTO>updateById(@PathVariable Long doctorId,
 			@RequestBody DoctorRequestDTO doctorDto) {
 		if(!docService.exists(doctorId)) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -61,7 +61,7 @@ public class DoctorController {
 	}
 	
 	@DeleteMapping(path = "/delete/{doctorId}")
-	public ResponseEntity<DoctorRequestDTO>deleteDoctor(@PathVariable Integer doctorId){
+	public ResponseEntity<DoctorRequestDTO>deleteDoctor(@PathVariable Long doctorId){
 		docService.deleteDoc(doctorId);
 		return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 	}
